refactor(node-utils): host the shared package.json read helpers
The `readJson` / `readPackageJson` helpers and the `PackageJson` type were only
reachable inside `@trezor/requirements`, yet they are generic node primitives.
Move them into `@trezor/node-utils` (a package usable across the repo) so any
node-context consumer can share them, and adopt them from requirements.

- node-utils: add `readJson<T>` and `readPackageJson<T = PackageJson>` + the
  `PackageJson` type. Purely additive — existing consumers (suite-desktop-core,
  transport-bridge, coinjoin, …) are untouched.
- requirements: drop the local copies from `workspaces.ts` / `fileSystem.ts`,
  depend on `@trezor/node-utils`, and import the helpers from there.
- firmware-releases: the concrete cross-cutting payoff — its hand-rolled
  `JSON.parse(readFileSync(...))` now uses the shared `readJson`.

Behavior-preserving: package type-check, the full requirements suite,
`requirements:verify` (byte-for-byte connect-public-dependencies snapshots +
firmware-release-version-monotonicity), lint and depcheck all stay green.
