+++ packages/requirements/package.json
@@ -14,6 +14,7 @@
         "requirements:fix": "tsx src/run.ts fix"
     },
     "dependencies": {
+        "@trezor/node-utils": "workspace:*",
         "@trezor/utils": "workspace:*",
         "chalk": "^5.6.2",
         "semver": "^7.7.1"
