+++ packages/requirements/src/workspaces.ts
@@ -1,7 +1,5 @@
 import { execFileSync } from 'node:child_process';
-import { join, resolve } from 'node:path';
-
-import { readJson } from './fileSystem';
+import { resolve } from 'node:path';
 
 type YarnWorkspaceInfo = {
     readonly name: string;
@@ -68,28 +66,3 @@ export const getWorkspaceDirectoryMap = (repoRoot: string): ReadonlyMap<string, 
 
     return map;
 };
-
-/**
- * Structural view of a `package.json`. Only the fields the requirements actually
- * read are typed explicitly; the trailing index signature keeps arbitrary
- * metadata (`repository`, `bugs`, `author`, …) accessible as `unknown` without
- * every consumer redeclaring its own subset.
- */
-export type PackageJson = {
-    readonly name?: string;
-    readonly version?: string;
-    readonly private?: boolean;
-    readonly type?: string;
-    readonly scripts?: Record<string, string | undefined>;
-    readonly workspaces?: { readonly packages?: ReadonlyArray<string> } | ReadonlyArray<string>;
-    readonly resolutions?: Record<string, string>;
-    readonly dependencies?: Record<string, string>;
-    readonly devDependencies?: Record<string, string>;
-    readonly optionalDependencies?: Record<string, string>;
-    readonly peerDependencies?: Record<string, string>;
-    readonly peerDependenciesMeta?: Record<string, { readonly optional?: boolean }>;
-    readonly [field: string]: unknown;
-};
-
-export const readPackageJson = <T = PackageJson>(workspaceDir: string): T =>
-    readJson<T>(join(workspaceDir, 'package.json'));
