+++ packages/requirements/tsconfig.json
@@ -1,5 +1,8 @@
 {
     "extends": "../../tsconfig.base.json",
     "compilerOptions": { "outDir": "libDev" },
-    "references": [{ "path": "../utils" }]
+    "references": [
+        { "path": "../node-utils" },
+        { "path": "../utils" }
+    ]
 }
